requirements.txt
fastapi
uvicorn
ollama
openai
python-multipart
python-dotenv
mysql-connector-python
